pyproject.toml
[project]
name = "echo-backend"
version = "0.1.0"
description = "..."
authors = [
    {name = "sebastianLedzianowski",email = "136706736+sebastianLedzianowski@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "uvicorn (>=0.35.0,<0.36.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "fastapi-mail (>=1.5.0,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "cryptography (>=45.0.6,<46.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "alembic (>=1.16.4,<2.0.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "httpie (>=3.2.4,<4.0.0)",
    "pytest-asyncio (>=1.1.0,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
